refactor: extract physical ai family
