+++ src/archetype/physical_ai/__init__.py
@@ -1,0 +1,9 @@
+# Copyright 2026 Vangelis Technologies Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Physical-AI components, processors, and external execution boundaries.
+
+This family owns reusable simulation and policy state. Application workflows
+that create worlds, run episodes, query ledgers, or grade outcomes live in the
+application layer and consume these definitions.
+"""
