+++ tests/physical_ai/test_boundary_helper.py
@@ -26,7 +26,7 @@ from __future__ import annotations
 import pytest
 from daft import Series
 
-from archetype.experiments.boundary import series_to_rows, unpack_struct
+from archetype.physical_ai.boundary import series_to_rows, unpack_struct
 
 # ---------------------------------------------------------------------------
 # series_to_rows
