+++ tests/physical_ai/test_evaluation_integration.py
@@ -32,7 +32,7 @@ from archetype.app.container import ServiceContainer
 from archetype.app.gateway.auth.guard import reset_daily_tokens, reset_tick_counters
 from archetype.app.models import EpisodeConfig
 from archetype.core.config import StorageConfig, WorldConfig
-from archetype.experiments.manipulation import (
+from archetype.physical_ai.manipulation import (
     ACTION_DIM,
     EnvStepProcessor,
     ManipAction,
