+++ tests/physical_ai/test_policy_loop.py
@@ -20,7 +20,7 @@ import pytest
 from archetype.core.aio import AsyncSystem
 from archetype.core.component import Component
 from archetype.core.config import RunConfig, StorageConfig, WorldConfig
-from archetype.experiments.manipulation import (
+from archetype.physical_ai.manipulation import (
     ACTION_DIM,
     EnvStepProcessor,
     ManipAction,
@@ -30,7 +30,7 @@ from archetype.experiments.manipulation import (
     ManipTask,
     ScriptedReachEnv,
 )
-from archetype.experiments.policy import PolicyActionProcessor, ScriptedReachPolicy
+from archetype.physical_ai.policy import PolicyActionProcessor, ScriptedReachPolicy
 from tests.conftest import make_world_service
 
 SIG = (ManipAction, ManipProprio, ManipStatus, ManipTask)
